Added args, began code generation.
The number of different arguments the compiler takes has increased, and the main function reflects those updates.
A compile function has been added, where currently it detects which assembler to use, using gas only when NASM is not present.
Lots of boilerplate assembly code is set to be generated.

TODO: There is currently no facility in the parsing stage to save the directives needed for customizing the Turing machine environment, so facilities must be added to compile correctly.