Build system organized, did other cleanup.
A simple Makefile was written to be used to compile the compiler, but
can be organized into a more friendly build system. Since this was the
first time the code was actually compiled, many small syntactic errors
were fixed. After some limited test runs, it appears that the logic thus
far acts as expected, including the fork-exec which check for NASM,
which is sexy. Also, some aesthetic changes were made to the output
messages, and messages were added.