Tape initialized, current code verified.
We can now proceed with writing the meat of the NASM syntax assembly, For now, gas
syntax is going to be considered a deferred cause, and will be ignored. The current
state of the compiler allocates a tape, loads input onto the tape if needed, prints
the content of the tape, and exits cleanly.